% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparePE.R
\name{Compare}
\alias{Compare}
\title{Compare two potential evapotranspiration methods}
\usage{
Compare(Method1, Method2)
}
\arguments{
\item{Method1}{A vector, 1-column matrix or data frame with potential evapotranspiration
calculated from Method1.}

\item{Method2}{A vector, 1-column matrix or data frame with potential evapotranspiration
calculated from Method2.}
}
\value{
a \code{list}, which contains:

\itemize{
\item Absolute mean error (AME),
\item square root of the mean squared error (RMSE),
\item Willmott's indices of agreement:
\itemize{
\item original (dorig),
\item modified (dmod) and
\item refined (dref)
}
\item Pearson determination coefficient (R2), and
}
}
\description{
Calculates measures of accuracy and agreement.
}
\examples{
data(DataForCWB)
Tavg <- DataForCWB[,2]
Tmax <- DataForCWB[,3]
Tmin <- DataForCWB[,4]
Rn <- DataForCWB[,6]
WS <- DataForCWB[,7]
RH <- DataForCWB[,7]
G <- DataForCWB[,9]
Method1 <- PE_PM(Tavg, Tmax, Tmin, Rn, RH, WS,G)
Method2 <- PE_PT(Tavg, Rn,G)
Comp <- Compare(Method1=Method1, Method2=Method2)
Comp
}
