% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropWaterBalance-package.R
\docType{package}
\name{CropWaterBalance-package}
\alias{CropWaterBalance}
\alias{CropWaterBalance-package}
\title{CropWaterBalance: Climate Water Balance for Irrigation Purposes}
\description{
Calculates daily climate water balance for irrigation purposes. The package also calculates the reference 'evapotranspiration' using three methods, 'Penman and Monteith', 'Priestley and Taylor', or 'Hargreaves and Samani'. Users may specify a management allowed depletion ('MAD'), which is used by the package to suggest when irrigate. It also allows the use of crop and water stress coefficients.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gabrielblain/CropWaterBalance}
  \item Report bugs at \url{https://github.com/gabrielblain/CropWaterBalance/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Constantino Blain \email{gabriel.blain@sp.gov.br} (\href{https://orcid.org/0000-0001-8832-7734}{ORCID})

Authors:
\itemize{
  \item Graciela R. Sobierajski (\href{https://orcid.org/0000-0002-7211-9268}{ORCID})
  \item Regina C. Matos Pires (\href{https://orcid.org/0000-0003-4200-7094}{ORCID})
  \item Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
  \item Leticia L. Martins (\href{https://orcid.org/0000-0002-0299-3005}{ORCID})
}

}
\keyword{internal}
