% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Descriptive.R
\name{Descriptive}
\alias{Descriptive}
\title{Descriptive Statistics for Weather Variables}
\usage{
Descriptive(Sample)
}
\arguments{
\item{Sample}{A vector, 1-column matrix or data frame with rainfall, evapotranspiration,
or other variable.}
}
\value{
A \code{dataframe} with:

\itemize{
\item sample mean (Avg),
\item sample median (Med),
\item sample standard variation (SD)
\item sample standard Error (SE)
\item maximum value (MaxValue)
\item minimum value (MinValue)
\item frequency of zeros (FreqZero\%)
}
}
\description{
Calculates descriptive statistics for rainfall, evapotranspiration, or other
variables.
}
\examples{
data(DataForCWB)
Rain <- DataForCWB[,10]
Descriptive(Sample = Rain)
}
