% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETr_HS.R
\name{ETr_HS}
\alias{ETr_HS}
\title{Reference 'evapotranspiration'}
\usage{
ETr_HS(Ra, Tavg, Tmax, Tmin)
}
\arguments{
\item{Ra}{Extraterrestrial solar radiation in \acronym{MJ m-2 day-1}.}

\item{Tavg}{A vector, 1-column matrix or data frame with daily average air temperature.}

\item{Tmax}{A vector, 1-column matrix or data frame with daily Maximum air temperature in Celsius degrees.}

\item{Tmin}{A vector, 1-column matrix or data frame with daily Minimum air temperature in Celsius degrees.}
}
\value{
Daily potential evapotranspiration values.
}
\description{
Daily reference 'evapotranspiration' using 'Hargreaves' and 'Samani' method.
}
\examples{
data(DataForCWB)
Tavg <- DataForCWB[,2]
Tmax <- DataForCWB[,3]
Tmin <- DataForCWB[,4]
Ra <- DataForCWB[,5]
ETr_HS(Ra, Tavg, Tmax, Tmin)
}
