% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETr_PM.R
\name{ETr_PM}
\alias{ETr_PM}
\title{Reference 'evapotranspiration'}
\usage{
ETr_PM(Tavg, Tmax, Tmin, Rn, RH, WS, G = NULL)
}
\arguments{
\item{Tavg}{A vector, 1-column matrix or data frame with daily average air temperature.}

\item{Tmax}{A vector, 1-column matrix or data frame with daily Maximum air temperature in Celsius degrees.}

\item{Tmin}{A vector, 1-column matrix or data frame with daily Minimum air temperature in Celsius degrees.}

\item{Rn}{A vector, 1-column matrix or data frame with daily net radiation in \acronym{MJ m-2 day-1}.}

\item{RH}{A vector, 1-column matrix or data frame with daily Relative Humidity  in \%.}

\item{WS}{A vector, 1-column matrix or data frame with daily Wind speed in \acronym{m s-1}.}

\item{G}{Optional. A vector, 1-column matrix or data frame with daily soil Heat flux in \acronym{MJ m-2 day-1}.
May be provided by \code{\link{Soil_Heat_Flux}}}
}
\value{
Daily potential evapotranspiration values (Penman & Monteith \acronym{FAO-1998}).
}
\description{
Daily reference 'evapotranspiration' using 'Penman' and 'Monteith' method.
}
\examples{
data(DataForCWB)
Tavg <- DataForCWB[,2]
Tmax <- DataForCWB[,3]
Tmin <- DataForCWB[,4]
Rn <- DataForCWB[,6]
WS <- DataForCWB[,7]
RH <- DataForCWB[,8]
G <- DataForCWB[,9]
ETr_PM(Tavg, Tmax, Tmin, Rn, RH, WS,G)
}
