% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETr_PT.R
\name{ETr_PT}
\alias{ETr_PT}
\title{Reference 'evapotranspiration'}
\usage{
ETr_PT(Tavg, Rn, G = NULL)
}
\arguments{
\item{Tavg}{A vector, 1-column matrix or data frame with daily average air temperature.}

\item{Rn}{A vector, 1-column matrix or data frame with daily net radiation in \acronym{MJ m-2 day-1}}

\item{G}{Optional. A vector, 1-column matrix or data frame with daily soil Heat flux in \acronym{MJ m-2 day-1}.
May be provided by \code{\link{Soil_Heat_Flux}}}
}
\value{
Daily potential evapotranspiration values.
}
\description{
Daily reference 'evapotranspiration' using 'Priestley' and 'Taylor' method.
}
\examples{
data(DataForCWB)
Tavg <- DataForCWB[,2]
Rn <- DataForCWB[,4]
G <- DataForCWB[,9]
ETr_PT(Tavg, Rn,G)
}
