% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ks.R
\name{Ks}
\alias{Ks}
\title{Water stress coefficient}
\usage{
Ks(AWC, MAD = NULL, Drz, SoilWaterDeficit)
}
\arguments{
\item{AWC}{Vector, 1-column matrix or data frame defining the available water capacity of the soil, that is:
the amount of water between field capacity and permanent wilting point in centimeter of water per centimeter of soil.}

\item{MAD}{Vector, 1-column matrix or data frame defining the management allowed depletion.
Varies between 0.1 and 1, default = 0.8.}

\item{Drz}{Vector, 1-column matrix or data frame defining the root zone depth in centimeters.}

\item{SoilWaterDeficit}{Vector, 1-column matrix or data frame with soil water deficit in millimeters.
It is provided by the \code{\link{CWB}} function.}
}
\value{
Water stress coefficient \acronym{Ks}.
}
\description{
Calculates the water stress coefficient (Ks).
}
\examples{
data(DataForCWB)
Tavg <- DataForCWB[,2]
Tmax <- DataForCWB[,3]
Tmin <- DataForCWB[,4]
Rn <- DataForCWB[,6]
WS <- DataForCWB[,7]
RH <- DataForCWB[,8]
G <- DataForCWB[,9]
ETr <- ETr_PM(Tavg, Tmax, Tmin, Rn, RH, WS,G)
Rain <- DataForCWB[,10]
Drz <- DataForCWB[,11]
AWC <- DataForCWB[,12]
MAD <- DataForCWB[,13]
CWB.been <- CWB(Rain, ETr, AWC,Drz, start.date = "2023-11-23")
D <- CWB.been$SoilWaterDeficit
Ks(AWC = AWC,MAD = MAD, Drz = Drz, SoilWaterDeficit = D)
}
