% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_HS.R
\name{PE_HS}
\alias{PE_HS}
\title{PE_HS}
\usage{
PE_HS(Ra, Tavg, Tmax, Tmin)
}
\arguments{
\item{Ra}{Extraterrestrial solar radiation in \acronym{MJ m-2 day-1}.}

\item{Tavg}{A vector, 1-column matrix or data frame with daily average air temperature.}

\item{Tmax}{A vector, 1-column matrix or data frame with daily Maximum air temperature in Celsius degrees.}

\item{Tmin}{A vector, 1-column matrix or data frame with daily Minimum air temperature in Celsius degrees.}
}
\value{
Daily potential evapotranspiration values.
}
\description{
PE_HS
}
\examples{
data(DataForCWB)
Tavg <- DataForCWB[,2]
Tmax <- DataForCWB[,3]
Tmin <- DataForCWB[,4]
Ra <- DataForCWB[,5]
PE_HS(Ra, Tavg, Tmax, Tmin)
}
