% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Water_Stress_Coef.R
\name{Water_Stress_Coef}
\alias{Water_Stress_Coef}
\title{Water stress coefficient}
\usage{
Water_Stress_Coef(AWC, MAD, Drz, SoilWaterDeficit)
}
\arguments{
\item{AWC}{Vector, 1-column matrix or data frame defining the available water capacity of the soil, that is:
the amount of water between field capacity and permanent wilting point in millimeters of water
per centimeter of soil.}

\item{MAD}{Vector, 1-column matrix or data frame defining the management allowed depletion.
Varies between 0 and 1, default is 0.3.}

\item{Drz}{Vector, 1-column matrix or data frame defining the root zone depth in centimeters.}

\item{SoilWaterDeficit}{Vector, 1-column matrix or data frame with soil water deficit in millimeters.
It may be provided by the \code{\link{CWB}} function.}
}
\value{
Water stress coefficient (Ks).
}
\description{
Calculates the water stress coefficient (Ks).
}
\examples{
data(DataForCWB)
Tavg <- DataForCWB[,2]
Tmax <- DataForCWB[,3]
Tmin <- DataForCWB[,4]
Rn <- DataForCWB[,6]
WS <- DataForCWB[,7]
RH <- DataForCWB[,8]
G <- DataForCWB[,9]
ET0 <- ET0_PM(Tavg, Tmax, Tmin, Rn, RH, WS,G)
Rain <- DataForCWB[,10]
Drz <- DataForCWB[,11]
AWC <- DataForCWB[,12]
MAD <- DataForCWB[,13]
Kc <- DataForCWB[,14]
Ks <- DataForCWB[,15]
Irrig <- DataForCWB[,16]
CWB.been <- CWB(Rain=Rain, ET0=ET0, AWC=AWC, Drz=Drz,
    Kc=Kc, Ks=Ks, Irrig=Irrig, MAD=MAD, start.date = "2023-11-23")
D <- CWB.been$SoilWaterDeficit
Water_Stress_Coef(AWC = AWC,MAD = MAD, Drz = Drz, SoilWaterDeficit = D)
}
