% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWB.R
\name{CWB}
\alias{CWB}
\title{CWB}
\usage{
CWB(Rain, PE, AWC, STinit = AWC, Kc = NULL, Ks = NULL, Irrig = NULL)
}
\arguments{
\item{Rain}{A vector, 1-column matrix or data frame with daily rainfall totals in millimeters.}

\item{PE}{A vector, 1-column matrix or data frame with daily potential evapotranspiration in millimeters.}

\item{AWC}{A single number defining the available water capacity in the root zone.}

\item{STinit}{A single number defining the water capacity in the root zone at the beginning
of the water balance accounting. Default is \dQuote{STinit=AWC}}

\item{Kc}{A single number defining the crop coefficient. Default is 1.}

\item{Ks}{A single number defining the water stress coefficient. Default is 1.}

\item{Irrig}{A vector, 1-column matrix or data frame with net irrigation amount infiltrated into the soil
for the current day in millimeters. Default is 0.}
}
\value{
Daily potential evapotranspiration values (Penman & Monteith \acronym{FAO-1998}).
}
\description{
CWB
}
\examples{
data(DataForCWB)
Tavg <- DataForCWB[,2]
Tmax <- DataForCWB[,3]
Tmin <- DataForCWB[,4]
Rn <- DataForCWB[,6]
WS <- DataForCWB[,7]
RH <- DataForCWB[,7]
G <- DataForCWB[,9]
PE <- PE_PM(Tavg, Tmax, Tmin, Rn, RH, WS,G)
Rain <- DataForCWB[,10]
CWB(Rain, PE, AWC=50)
}
