% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWB.R
\name{CWB}
\alias{CWB}
\title{Crop Water Balance Accounting}
\usage{
CWB(
  Rain,
  ETr,
  AWC,
  Drz,
  Kc = NULL,
  Ks = NULL,
  Irrig = NULL,
  MAD = NULL,
  start.date
)
}
\arguments{
\item{Rain}{Vector, 1-column matrix or data frame with daily rainfall totals in millimeters.}

\item{ETr}{Vector, 1-column matrix or data frame with daily reference evapotranspiration in millimeters.}

\item{AWC}{Vector, 1-column matrix or data frame defining the available water capacity of the soil, that is:
the amount of water between field capacity and permanent wilting point in millimeter of water per centimeter of soil.}

\item{Drz}{Vector, 1-column matrix or data frame defining the root zone depth in centimeters.}

\item{Kc}{Single number defining the crop coefficient. Default is 1.}

\item{Ks}{Single number defining the water stress coefficient. Default is 1.}

\item{Irrig}{Vector, 1-column matrix or data frame with net irrigation amount infiltrated into the soil
for the current day in millimeters. Default is 0.}

\item{MAD}{Vector, 1-column matrix or data frame defining the management allowed depletion.
Varies between 0 and 1, default = 0.3.}

\item{start.date}{date at which the accounting should start. Formats:
\dQuote{YYYY-MM-DD}, \dQuote{YYYY/MM/DD}.}
}
\value{
Water Balance Accounting, including the soil water defict.
}
\description{
Calculates several parameters of the crop water balance.
It also suggests when irrigate.
}
\examples{
data(DataForCWB)
Tavg <- DataForCWB[,2]
Tmax <- DataForCWB[,3]
Tmin <- DataForCWB[,4]
Rn <- DataForCWB[,6]
WS <- DataForCWB[,7]
RH <- DataForCWB[,8]
G <- DataForCWB[,9]
ETr <- ETr_PM(Tavg, Tmax, Tmin, Rn, RH, WS,G)
Rain <- DataForCWB[,10]
Drz <- DataForCWB[,11]
AWC <- DataForCWB[,12]
MAD <- DataForCWB[,13]
CWB(Rain, ETr, AWC,Drz,MAD,start.date = "2023-11-23")
}
