% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataForCWB.R
\docType{data}
\name{DataForCWB}
\alias{DataForCWB}
\title{Data for Water Balance Accounting}
\format{
An object of class \code{data.frame} with 129 rows and 13 columns.
}
\usage{
DataForCWB
}
\description{
Daily meteorological data from a weather station in Campinas, Brazil.
Belongs to the Agronomic Institute \acronym{IAC}
}
\details{
@format ## \code{DataForCWB}
A data frame with 13 columns and 129 rows:
\describe{
\item{date}{date}
\item{tmed}{Average air temperature in Celsius degrees}
\item{tmax}{Maximum air temperature in Celsius degrees}
\item{tmin}{Minimum air temperature in Celsius degrees}
\item{Ra}{Extraterrestrial solar radiation in \acronym{MJ m-2 day-1}}
\item{Rn}{Net radiation in \acronym{MJ m-2 day-1}}
\item{W}{Wind speed in \acronym{m s-1}}
\item{RH}{Relative Humidity  in \%}
\item{G}{Soil Heat Flux  in \acronym{MJ m-2 day-1}}
\item{Rain}{Rain in millimeters}
\item{Drz}{Depth of the root zone in centimeters}
\item{AWC}{available water capacity (amount of water between field capacity and permanent wilting point)
in millimeter of water per centimeter of soil}
\item{MAD}{management allowed depletion (between 0 and 1)}
}
@source \url{https://clima.iac.sp.gov.br/}
}
\keyword{datasets}
