% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil_Heat_Flux.R
\name{Soil_Heat_Flux}
\alias{Soil_Heat_Flux}
\title{Soil_Heat_Flux}
\usage{
Soil_Heat_Flux(Tavg)
}
\arguments{
\item{Tavg}{A vector, 1-column matrix or data frame with daily average air temperature.}
}
\value{
Daily values for soil Heat flux in \acronym{MJ/m^2/day}.
}
\description{
Calculates the Soil Heat Flux
}
\examples{
data(DataForCWB)
Tavg <- DataForCWB[,2]
Soil_Heat_Flux(Tavg)
}
